tags: move to internal/tags package
